requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
exceptiongroup==1.2.0
fastapi==0.108.0
filelock==3.13.1
fonttools==4.47.0
fsspec==2023.12.2
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
imageio==2.33.1
importlib-resources==6.1.1
iniconfig==2.0.0
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.7.4
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
pluggy==1.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.1
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
starlette==0.32.0.post1
sympy==1.12
threadpoolctl==3.2.0
tifffile==2023.7.10
tomli==2.0.1
torch==2.1.2
torchvision==0.16.2
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.25.0
watchfiles==0.21.0
websockets==12.0
zipp==3.17.0
